test: migrate SearchSelect to vitest
